refactor: add action creator for loadNews
